@epwalsh 2020-02-24
This is great!!

Do you think it's worth having a test for the trainer? I'm thinking the test would just feed in a very small hash map of word counts to `BpeTrainer::train` so we can the resulting vocab and merges exactly. I could make a separate PR to master with that to master if you think it's worth it.