@n1t0 2020-02-24
Definitely! That's a great idea, feel free to make a PR!